Hope you like printlns (+HexGrid & persist fixes)
Lots of println statements. Fixed the offset stuff for HexGrid.java (i
think), Added an equal sign to the if-statement of line 38 in
CommandListXMLmodel.java because persist wasn't working without it.
Still, I don't completely understand the whole thing, so if this screws
something up, feel free to take the equal sign out...but then the build
stuff commands with not work.